Moves classes in good packages.
