+++ bundles/orbistoolbox/src/main/java/org/orbisgis/orbistoolbox/controller/execution/DataProcessingManager.java
@@ -17,7 +17,7 @@
  * For more information, please consult: <http://www.orbisgis.org/> or contact directly: info_at_orbisgis.org
  */
 
-package org.orbisgis.orbistoolbox.view.utils.dataProcessing;
+package org.orbisgis.orbistoolbox.controller.execution;
 
 import org.orbisgis.orbistoolbox.model.DataDescription;
 import org.orbisgis.orbistoolbox.model.DescriptionType;
