+++ bundles/orbistoolbox/src/main/java/org/orbisgis/orbistoolbox/controller/parser/EnumerationParser.java
@@ -19,7 +19,8 @@
 
 package org.orbisgis.orbistoolbox.controller.parser;
 
-import org.orbisgis.orbistoolbox.controller.processexecution.utils.FormatFactory;
+import org.orbisgis.orbistoolbox.controller.utils.FormatFactory;
+import org.orbisgis.orbistoolbox.controller.utils.ObjectAnnotationConverter;
 import org.orbisgis.orbistoolbox.model.*;
 import org.orbisgis.wpsgroovyapi.attributes.DescriptionTypeAttribute;
 import org.orbisgis.wpsgroovyapi.attributes.EnumerationAttribute;
