+++ bundles/orbistoolbox/src/main/java/org/orbisgis/orbistoolbox/controller/parser/RawDataParser.java
@@ -19,6 +19,7 @@
 
 package org.orbisgis.orbistoolbox.controller.parser;
 
+import org.orbisgis.orbistoolbox.controller.utils.ObjectAnnotationConverter;
 import org.orbisgis.orbistoolbox.model.*;
 import org.orbisgis.wpsgroovyapi.attributes.*;
 import org.slf4j.LoggerFactory;
