+++ bundles/orbistoolbox/src/main/java/org/orbisgis/orbistoolbox/controller/utils/FormatFactory.java
@@ -17,7 +17,7 @@
  * For more information, please consult: <http://www.orbisgis.org/> or contact directly: info_at_orbisgis.org
  */
 
-package org.orbisgis.orbistoolbox.controller.processexecution.utils;
+package org.orbisgis.orbistoolbox.controller.utils;
 
 import org.orbisgis.orbistoolbox.model.Format;
 import org.orbisgis.orbistoolbox.model.MalformedScriptException;
