+++ bundles/orbistoolbox/src/main/java/org/orbisgis/orbistoolbox/view/ToolBox.java
@@ -25,17 +25,16 @@ import org.h2gis.h2spatialapi.DriverFunction;
 import org.h2gis.h2spatialapi.EmptyProgressVisitor;
 import org.h2gis.utilities.JDBCUtilities;
 import org.h2gis.utilities.TableLocation;
-import org.orbisgis.commons.progress.SwingWorkerPM;
 import org.orbisgis.corejdbc.DataManager;
 import org.orbisgis.dbjobs.api.DriverFunctionContainer;
 import org.orbisgis.frameworkapi.CoreWorkspace;
-import org.orbisgis.orbistoolbox.controller.ProcessManager;
+import org.orbisgis.orbistoolbox.controller.process.ProcessManager;
 import org.orbisgis.orbistoolbox.model.DataType;
 import org.orbisgis.orbistoolbox.model.Process;
 import org.orbisgis.orbistoolbox.view.ui.ToolBoxPanel;
 import org.orbisgis.orbistoolbox.view.ui.dataui.DataUIManager;
 import org.orbisgis.orbistoolbox.view.utils.*;
-import org.orbisgis.orbistoolbox.view.utils.dataProcessing.DataProcessingManager;
+import org.orbisgis.orbistoolbox.controller.execution.DataProcessingManager;
 import org.orbisgis.orbistoolbox.view.utils.editor.log.LogEditableElement;
 import org.orbisgis.orbistoolbox.view.utils.editor.log.LogEditor;
 import org.orbisgis.orbistoolbox.view.utils.editor.process.ProcessEditableElement;
